Removed extra unused code and rearranged
