+++ BlockBlobReader/requirements.txt
@@ -10,3 +10,4 @@ azure-mgmt-authorization==4.0.0
 azure-graphrbac==0.61.1
 azure-identity==1.15.0
 sumologic-sdk==0.1.15
+azure-mgmt-loganalytics==12.0.0
