+++ BlockBlobReader/tests/baseblockblobtest.py
@@ -1,27 +1,16 @@
 import json
 import os
 import sys
-from sumologic import SumoLogic
-from azure.mgmt.resource import ResourceManagementClient
 
 sys.path.insert(0, '../../test_utils')
 from basetest import BaseTest
 
 
 class BaseBlockBlobTest(BaseTest):
-
-    def setUp(self):
-        self.create_credentials()
-        self.resource_client = ResourceManagementClient(self.azure_credential, self.subscription_id)
-        self.repo_name, self.branch_name = self.get_git_info()
-        self.sumologic_cli = SumoLogic(os.environ["SUMO_ACCESS_ID"], os.environ["SUMO_ACCESS_KEY"], self.api_endpoint(os.environ["SUMO_DEPLOYMENT"]))
-        self.collector_id = self.create_collector(self.collector_name)
-        self.sumo_source_id, self.sumo_endpoint_url = self.create_source(self.collector_id, self.source_name)
-        
+  
     def tearDown(self):
         if self.resource_group_exists(self.RESOURCE_GROUP_NAME):
             self.delete_resource_group()
-            self.delete_event_subscription()
         self.delete_container()
         self.delete_source(self.collector_id, self.sumo_source_id)
         self.delete_collector(self.collector_id)
