+++ EventHubs/tests/baseeventhubtest.py
@@ -2,13 +2,11 @@ import os
 import sys
 import json
 from datetime import timedelta
-from sumologic import SumoLogic
-from azure.mgmt.resource import ResourceManagementClient
 from azure.mgmt.eventhub import EventHubManagementClient
 from azure.eventhub import EventHubProducerClient
 from azure.eventhub.exceptions import EventHubError
 from azure.monitor.query import LogsQueryClient, LogsQueryStatus
-from azure.mgmt.loganalytics import LogAnalyticsManagementClient
+
 
 sys.path.insert(0, '../../test_utils')
 from basetest import BaseTest
@@ -16,33 +14,12 @@ from basetest import BaseTest
 
 class BaseEventHubTest(BaseTest):
 
-    def setUp(self):
-        self.create_credentials()
-        self.resource_client = ResourceManagementClient(self.azure_credential, 
-                                                        self.subscription_id)
-        self.repo_name, self.branch_name = self.get_git_info()
-        self.sumologic_cli = SumoLogic(os.environ["SUMO_ACCESS_ID"], os.environ["SUMO_ACCESS_KEY"], self.api_endpoint(os.environ["SUMO_DEPLOYMENT"]))
-        self.collector_id = self.create_collector(self.collector_name)
-        self.sumo_source_id, self.sumo_endpoint_url = self.create_source(self.collector_id, self.source_name)
-        
     def tearDown(self):
         if self.resource_group_exists(self.RESOURCE_GROUP_NAME):
             self.delete_resource_group()
         self.delete_source(self.collector_id, self.sumo_source_id)
         self.delete_collector(self.collector_id)
         self.sumologic_cli.session.close()
-
-    def get_resource_name(self, resprefix, restype):
-        for item in self.resource_client.resources.list_by_resource_group(self.RESOURCE_GROUP_NAME):
-            if (item.name.startswith(resprefix) and item.type == restype):
-                return item.name
-        raise Exception("%s Resource Not Found" % (resprefix))
-    
-    def get_resource(self, restype):
-        for item in self.resource_client.resources.list_by_resource_group(self.RESOURCE_GROUP_NAME):
-            if (item.type == restype):
-                return item
-        raise Exception("%s Resource Not Found" % (restype))
 
     def _parse_template(self):
         template_path = os.path.join(os.path.abspath('..'), 'src',
@@ -105,22 +82,6 @@ class BaseEventHubTest(BaseTest):
             print("Exception", e)
 
         return result
-    
-    def get_resources(self, resource_group_name):
-        return self.resource_client.resources.list_by_resource_group(resource_group_name)
-    
-    def get_Workspace_Id(self):
-        workspace = self.get_resource('microsoft.operationalinsights/workspaces')
-        client = LogAnalyticsManagementClient(
-            credential=self.azure_credential,
-            subscription_id=self.subscription_id,
-        )
-
-        response = client.workspaces.get(
-            resource_group_name=self.RESOURCE_GROUP_NAME,
-            workspace_name=workspace.name,
-        )
-        return response.customer_id
     
     def filter_logs(self, logs, key, value):
         return value in [d.get(key) for d in logs]
